CRUD correction front and back

+++ Soutenances/src/main/java/com/example/Soutenances/Repositories/SoutenancesRepository.java
@@ -7,8 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SoutenancesRepository extends JpaRepository<Soutenances, Long> {
+public interface SoutenancesRepository extends JpaRepository<Soutenances, Long> { // Correction : Changement de int à Long
     List<Soutenances> findByDepartementId(Long departementId);
-    boolean existsById(int id);
-
+    boolean existsById(Long id); // Correction : Changement de int à Long
 }
